CaseList: move to cases/
